requirements.txt
algoliasearch==3.0.0
APScheduler==3.10.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-analytics-data==0.18.12
google-api-core==2.20.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
numpy==2.1.1
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
ShopifyAPI==12.6.0
six==1.16.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
